+++ src/2021/9/run.ts
@@ -1,5 +1,4 @@
-import { getInput } from '../../lib/getInput';
-import { log } from '../../lib/log';
+import { getInput, log } from '../../lib';
 
 const input = getInput(2021, 9)
     .split('\n')
