+++ src/lib/index.ts
@@ -1,0 +1,3 @@
+export { getInput } from './getInput';
+export { isTruthy } from './isTruthy';
+export { log } from './log';
